refactor: extract cli file helper module
- Move managed-file and frontmatter helpers out of cli.ts
- Add colocated tests for merge/skip/abort and copy paths
- Keep CLI behavior unchanged while improving cohesion

Made-with: Cursor
