+++ src/cli.ts
@@ -16,6 +16,7 @@ import {
   resolveWorkspacePath,
   workspaceLabelFor
 } from './cli/context-resolution.ts';
+import { copySourceFile, parseFrontmatter, writeManagedFile } from './cli/file-helpers.ts';
 import { isRecord, validateWorkspaceOverride } from './cli/override-validation.ts';
 import { listWorkspaceDirs } from './cli/workspace-discovery.ts';
 import type {
@@ -37,8 +38,6 @@ import type {
   WorkspaceState
 } from './cli/types.ts';
 
-const AILIB_BLOCK_START = '<!-- ailib:start -->';
-const AILIB_BLOCK_END = '<!-- ailib:end -->';
 const CONFIG_FILE = 'ailib.config.json';
 const LOCAL_OVERRIDE_FILE = 'ailib.local.json';
 const LOCK_FILE = 'ailib.lock';
@@ -1296,75 +1295,6 @@ async function writeRootLock({
   }
 
   await fs.writeFile(path.join(rootDir, LOCK_FILE), `${JSON.stringify(lock, null, 2)}\n`, 'utf8');
-}
-
-async function writeManagedFile({
-  outPath,
-  rendered,
-  onConflict
-}: {
-  outPath: string;
-  rendered: string;
-  onConflict: string;
-}) {
-  await fs.mkdir(path.dirname(outPath), { recursive: true });
-
-  if (await exists(outPath)) {
-    if (onConflict === 'skip') return;
-    if (onConflict === 'abort') {
-      throw new Error(`Conflict detected for ${outPath}; rerun with --on-conflict=overwrite|merge|skip`);
-    }
-    if (onConflict === 'merge') {
-      const existing = await fs.readFile(outPath, 'utf8');
-      const withoutOld = existing.includes(AILIB_BLOCK_START)
-        ? `${existing.slice(0, existing.indexOf(AILIB_BLOCK_START)).trimEnd()}\n`
-        : `${existing.trimEnd()}\n`;
-      const merged = `${withoutOld}\n${AILIB_BLOCK_START}\n${rendered.trim()}\n${AILIB_BLOCK_END}\n`;
-      await fs.copyFile(outPath, `${outPath}.bak`);
-      await fs.writeFile(outPath, merged, 'utf8');
-      return;
-    }
-    await fs.copyFile(outPath, `${outPath}.bak`);
-  }
-
-  await fs.writeFile(outPath, `${rendered.trim()}\n`, 'utf8');
-}
-
-async function copySourceFile({
-  packageRoot,
-  sourceRel,
-  target
-}: {
-  packageRoot: string;
-  sourceRel: string;
-  target: string;
-}) {
-  const source = path.join(packageRoot, sourceRel);
-  ensure(await exists(source), `Missing module source: ${sourceRel}`);
-  await fs.mkdir(path.dirname(target), { recursive: true });
-  await fs.copyFile(source, target);
-}
-
-function parseFrontmatter(markdown: string): Record<string, string | string[]> | null {
-  const match = markdown.match(/^---\n([\s\S]*?)\n---\n/u);
-  if (!match) return null;
-  const fields: Record<string, string | string[]> = {};
-  for (const line of match[1].split('\n')) {
-    if (!line.trim() || line.trim().startsWith('#')) continue;
-    const idx = line.indexOf(':');
-    if (idx < 0) continue;
-    const key = line.slice(0, idx).trim();
-    let value: string | string[] = line.slice(idx + 1).trim();
-    if (value.startsWith('[') && value.endsWith(']')) {
-      value = value
-        .slice(1, -1)
-        .split(',')
-        .map((x) => x.trim())
-        .filter(Boolean);
-    }
-    fields[key] = value;
-  }
-  return fields;
 }
 
 function sha256(value: string) {
