+++ src/cli/file-helpers.test.ts
@@ -1,0 +1,60 @@
+import test from 'node:test';
+import assert from 'node:assert/strict';
+import fs from 'node:fs/promises';
+import os from 'node:os';
+import path from 'node:path';
+import { copySourceFile, parseFrontmatter, writeManagedFile } from './file-helpers.ts';
+
+async function tempDir() {
+  return fs.mkdtemp(path.join(os.tmpdir(), 'ailib-file-helpers-'));
+}
+
+test('parseFrontmatter parses key-value and list fields', () => {
+  const parsed = parseFrontmatter(
+    ['---', 'id: eslint', 'tags: [lint, style]', '# comment', 'language: typescript', '---', '# body'].join('\n')
+  );
+  assert.deepEqual(parsed, { id: 'eslint', tags: ['lint', 'style'], language: 'typescript' });
+});
+
+test('parseFrontmatter returns null when block missing', () => {
+  assert.equal(parseFrontmatter('# no frontmatter\n'), null);
+});
+
+test('writeManagedFile supports overwrite, skip, merge, and abort modes', async () => {
+  const root = await tempDir();
+  const target = path.join(root, 'rules.md');
+
+  await writeManagedFile({ outPath: target, rendered: 'first', onConflict: 'overwrite' });
+  assert.equal(await fs.readFile(target, 'utf8'), 'first\n');
+
+  await writeManagedFile({ outPath: target, rendered: 'second', onConflict: 'skip' });
+  assert.equal(await fs.readFile(target, 'utf8'), 'first\n');
+
+  await writeManagedFile({ outPath: target, rendered: 'merged-content', onConflict: 'merge' });
+  const merged = await fs.readFile(target, 'utf8');
+  assert.match(merged, /<!-- ailib:start -->/);
+  assert.match(merged, /merged-content/);
+  assert.equal(await fs.readFile(`${target}.bak`, 'utf8'), 'first\n');
+
+  await assert.rejects(
+    writeManagedFile({ outPath: target, rendered: 'ignored', onConflict: 'abort' }),
+    /Conflict detected/
+  );
+});
+
+test('copySourceFile copies from package source and validates existence', async () => {
+  const root = await tempDir();
+  const packageRoot = path.join(root, 'pkg');
+  const targetRoot = path.join(root, 'out');
+  await fs.mkdir(path.join(packageRoot, 'core'), { recursive: true });
+  await fs.writeFile(path.join(packageRoot, 'core', 'behavior.md'), '# behavior\n', 'utf8');
+
+  const target = path.join(targetRoot, 'behavior.md');
+  await copySourceFile({ packageRoot, sourceRel: 'core/behavior.md', target });
+  assert.equal(await fs.readFile(target, 'utf8'), '# behavior\n');
+
+  await assert.rejects(
+    copySourceFile({ packageRoot, sourceRel: 'core/missing.md', target: path.join(targetRoot, 'missing.md') }),
+    /Missing module source: core\/missing\.md/
+  );
+});
